backend/requirements/base.txt
# Bleeding edge Django
Django==4.1.2

# Configuration
django-environ==0.4.5
whitenoise==6.2.0
brotlipy==0.7.0

# Models
django-model-utils==3.1.1

# Rest apis
djangorestframework==3.11.2
django-filter==2.4.0
ujson==5.4.0

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.4

# Time zones support
pytz==2019.1

celery==5.2.2
# we use msgpack since it outperforms yaml and json de/serializers
msgpack_python==0.5.6

# Utils
django-extensions==3.2.1
django-watchman==1.3.0
django-debug-toolbar==3.7.0

# Admin
django-admin-list-filter-dropdown==1.0.2

# extra apps
django-comments-xtd==2.2.1
django-contrib-comments==1.9.1
django-simple-history==3.1.1
django-taggit==3.0.0
django-tagulous==1.3.3
